Coupled data to variables

+++ Sandbox/CookieClicker/src/MainCookie.java
@@ -4,7 +4,7 @@ public class MainCookie {
     String imageFile = "Sandbox/CookieClicker/cookie.png";
     int size = 400;
 
-    int clickCount;
+    Double clickCount;
     Ellipse2D boundingCircle;
 
 }
